requirements.txt
earthengine-api>=1.7
geemap
streamlit>=1.55
folium>=0.20
streamlit-folium>=0.26
plotly
pandas
psycopg2-binary
requests
beautifulsoup4
pyyaml
python-dotenv
pytest
